Fix some issues when generating packets
For some reason, I messed it really badly and robot packets were not
generated correctly

+++ src/Joysticks/Joysticks.h
@@ -42,7 +42,7 @@ struct GM_Joystick {
  */
 struct GM_Hat {
     int id;               /**< The raw ID of the hat */
-    short angle;          /**< The pressed angle of the hat */
+    int angle;            /**< The pressed angle of the hat */
     GM_Joystick joystick; /**< The joystick that the hat belongs to */
 };
